Move UpdateResult to POJO package.
