+++ src/main/java/com/rylinaux/plugman/pojo/UpdateResult.java
@@ -1,4 +1,4 @@
-package com.rylinaux.plugman.util;
+package com.rylinaux.plugman.pojo;
 
 /*
  * #%L
